arwspeedstr: arwspeedstr_update 97.0->97.9 (unit 97.77->98.46; #83c drop the (int) intermediate on the float->u8 alpha store; banked: FP whole-pair shift #82)

+++ src/main/dll/ARW/dll_02A2_arwspeedstr.c
@@ -70,7 +70,7 @@ void arwspeedstr_update(int obj) {
                 state->alpha = lbl_803E7108 * timeDelta + state->alpha;
                 if (state->alpha > *(f32 *)&lbl_803E710C)
                     state->alpha = lbl_803E710C;
-                ((GameObject *)obj)->anim.alpha = (int)state->alpha;
+                ((GameObject *)obj)->anim.alpha = state->alpha;
             }
         }
     }
